requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
openpyxl==3.1.2
reportlab==4.0.7
requests==2.31.0
python-dateutil==2.8.2


